Change how streams are parsed. Also change how chats are opened. (#165)
* Change how streams are parsed. Also change how chats are opened.

Before queries, parameters and fragments were removed. This broke some links to chat. Now the url is parsed properly on input, only the scheme and net location are cast to lower while the rest of the url is intact.

Youtube Live chat option  now also opens directly to the chat window.

* Travis.

* Remove unnecessary lower() on scheme.

* Change scheme to parsed_url.scheme.